Fixed non matching fields. Added first empty EJB Stateless service.

+++ webapp/src/main/java/org/coinexplorer/db/abe/entities/Datadir.java
@@ -11,7 +11,7 @@ public class Datadir implements Serializable {
 
 	@Id
 	@Column(name="datadir_id")
-	private long datadirId;
+	private BigDecimal datadirId;
 
 	@Column(name="blkfile_number")
 	private BigDecimal blkfileNumber;
@@ -27,7 +27,7 @@ public class Datadir implements Serializable {
 	public Datadir() {
 	}
 
-	public long getDatadirId() {
+	public BigDecimal getDatadirId() {
 		return this.datadirId;
 	}
